+++ source/extensions/isaacsim.sensors.physics/plugins/isaacsim.sensors.physics/contact_sensor/ContactSensor.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "../core/BaseSensorComponent.h"
+#include "../core/IsaacSensorComponent.h"
 #include "ContactManager.h"
 #include "IPhysicsSensor.h"