+++ source/extensions/isaacsim.sensors.physics/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [0.3.12] - 2025-03-17
+### Changed
+- Cleanup and rename BridgeApplication to PrimManager for clarity
+
+
 ## [0.3.11] - 2025-03-11
 ### Changed
 - Switch asset root for tests to internal nucleus