+++ source/extensions/isaacsim.sensors.physx/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [2.2.10] - 2025-03-17
+### Changed
+- Cleanup and rename BridgeApplication to PrimManager for clarity
+
+
 ## [2.2.9] - 2025-03-11
 ### Changed
 - Switch asset root for tests to internal nucleus